perf: 优化ldap配置
