+++ src/layout/components/GenericCreateUpdateForm/index.vue
@@ -37,6 +37,10 @@ export default {
       type: Object,
       default: () => ({})
     },
+    afterGetFormValue: {
+      type: Function,
+      default: (value) => value
+    },
     // 提交前，清理form的值
     cleanFormValue: {
       type: Function,
@@ -61,13 +65,14 @@ export default {
         return this.$t('common.createSuccessMsg')
       }
     },
-    // 更新成功的msg
+    // 保存成功，继续添加的msg
     saveSuccessContinueMsg: {
       type: String,
       default: function() {
         return this.$t('common.saveSuccessContinueMsg')
       }
     },
+    // 更新成功的msg
     updateSuccessMsg: {
       type: String,
       default: function() {
@@ -93,7 +98,9 @@ export default {
     objectDetailRoute: {
       type: Object,
       default: function() {
-        const routeName = this.$route.name.replace('Update', 'Detail').replace('Create', 'Detail')
+        const routeName = this.$route.name
+          .replace('Update', 'Detail')
+          .replace('Create', 'Detail')
         return { name: routeName }
       }
     },
@@ -233,7 +240,8 @@ export default {
     try {
       const values = await this.getFormValue()
       this.$log.debug('Final object is: ', values)
-      this.form = Object.assign(this.form, values)
+      const formValue = Object.assign(this.form, values)
+      this.form = this.afterGetFormValue(formValue)
     } finally {
       this.loading = false
     }
